refactor: move component tests to __tests__ directory
