+++ rs-react-app/src/__tests__/Card.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { Card } from './Card.tsx';
+import { Card } from '../components/Card.tsx';
 import type { Character } from '../api/rickAndMorty.ts';
 
 describe('Card', () => {
